move block class to com.example.examplemod.blocks
